rename: side menu item 위치 변경
